requirements.txt
python-telegram-bot==20.6
aiohttp==3.9.1
beautifulsoup4==4.12.2
selenium==4.16.0
aiosqlite==0.19.0
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0
requests==2.31.0
asyncio==3.4.3